requirements.txt
arrow==0.15.6
asgiref==3.2.10
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.8.2
begins==0.9
binaryornot==0.4.4
boltons==20.0.0
bunch==1.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cookiecutter==1.7.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
Cython==0.29.19
defusedxml==0.6.0
Django==3.0.8
django-admin-search==0.3.5
django-allauth==0.42.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.3.0
django-obfuscate==0.1.2
django-phonenumber-field==4.0.0
django-request-logging==0.7.1
django-rest-auth==0.9.5
django-sendgrid-v5==0.8.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
drf-yasg==1.17.1
drf-yasg-json-api==0.7.1
easycython==1.0.7
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jinja2-time==0.2.0
jsonschema==3.2.0
MarkupSafe==1.1.1
mock==4.0.2
numpy==1.18.4
oauthlib==3.1.0
packaging==20.4
pendulum==2.0.5
phonenumbers==8.12.6
postgres==3.0.0
poyo==0.5.0
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pyarmor==6.2.1
pyclean==1.0.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-http-client==3.2.7
python-slugify==4.0.0
python3-openid==3.2.0
pytz==2020.1
pytzdata==2019.3
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scons==3.1.2
sendgrid==6.4.4
setuptools-git==1.2
simple-configparser==0.1.4.0
six==1.15.0
soupsieve==2.0.1
sourcedefender==5.0.12
spark-parser==1.8.9
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
swagger-spec-validator==2.7.3
templated-mail==0.4.0.2
text-unidecode==1.3
TgCrypto==1.2.0
uncompyle6==3.7.0
unpyclib==0.8.1
uritemplate==3.0.1
urllib3==1.25.9
webencodings==0.5.1
xdis==4.6.0
yagmail==0.11.224
zipp==3.1.0
google-auth==1.20.0